requirements.txt
flask
sqlalchemy
flask-wtf
email_validator
flask-sqlalchemy
python-dotenv
flask-bcrypt
flask-login